Moved function to static class for better readability
